fix: remove redundant type cast

+++ src/main/java/de/clearit/kindergarten/appliance/DoubleToStringConverter.java
@@ -21,7 +21,7 @@ public class DoubleToStringConverter extends AbstractConverter {
   public Object convertFromSubject(Object subjectValue) {
     String result = "";
     if (subjectValue != null) {
-      result = ((Double) subjectValue).toString();
+      result = subjectValue.toString();
     }
     return result;
   }
